requirements.txt
Django==4.1.1
djangorestframework==3.13.1
Flask==2.2.2
matplotlib==3.6.0
mysql==0.0.3
mysql-connector==2.2.9
mysqlclient==2.1.1
numpy==1.23.3
openai==0.23.0
opencv-python==4.6.0.66
pandas==1.4.4
pandas-stubs==1.4.4.220912
Pillow==9.2.0
postgres==4.0
preshed==3.0.7
protobuf==4.21.6
psycopg2-binary==2.9.3
psycopg2-pool==1.1
scikit-learn==1.1.2
scipy==1.9.1
python-dotenv==0.21.0
gunicorn==20.1.0
beautifulsoup4==4.11.1
certifi==2022.9.14
charset-normalizer==2.1.1
fpdf==1.7.2
idna==3.4
lxml==4.9.1
requests==2.28.1
soupsieve==2.3.2.post1
urllib3==1.26.12
